+++ src/routes.js
@@ -2,6 +2,7 @@ import { lazy } from "react";
 
 const CGraph = lazy(() => import("views/graph.jsx"));
 const Company = lazy(() => import("views/Index.jsx"));
+const Skill = lazy(() => import("views/Skill.jsx"));
 
 const routes = [
   {
@@ -14,6 +15,12 @@ const routes = [
     path: "/graph",
     name: "Graph",
     component: CGraph,
+    layout: "/admin",
+  },
+  {
+    path: "/skill",
+    name: "Skill",
+    component: Skill,
     layout: "/admin",
   },
 ];